Reorganizar funciones ajenas a Io
